src/test/java/codesquad/webserver/processor/HttpRequestParserTest.java
package codesquad.webserver.processor;

import codesquad.webserver.exception.BadRequestException;
import codesquad.webserver.http.HttpMethod;
import codesquad.webserver.http.HttpRequest;
import codesquad.webserver.http.HttpVersion;
import codesquad.webserver.http.Path;
import codesquad.webserver.http.header.HeaderConstants;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.io.ByteArrayInputStream;
import java.io.IOException;
import java.io.InputStream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;

class HttpRequestParserTest {

    private final HttpRequestParser httpRequestParser = new HttpRequestParser();

    @DisplayName("parseRequest: MultiPart 요청이 적절하게 파싱되는지 확인")
    @Test
    void parseMultiPartRequest() throws IOException {
        // given
        String request = "POST /upload HTTP/1.1\r\n" +
                "Host: localhost\r\n" +
                "Content-Type: multipart/form-data; boundary=----WebKitFormBoundary7MA4YWxkTrZu0gW\r\n" +
                "Content-Length: 327\r\n" +
                "\r\n" +
                "------WebKitFormBoundary7MA4YWxkTrZu0gW\r\n" +
                "Content-Disposition: form-data; name=\"content\"\r\n" +
                "\r\n" +
                "This is the content\r\n" +
                "------WebKitFormBoundary7MA4YWxkTrZu0gW\r\n" +
                "Content-Disposition: form-data; name=\"image\"; filename=\"example.png\"\r\n" +
                "Content-Type: image/png\r\n" +
                "\r\n" +
                "<binary data of example.png>\r\n" +
                "------WebKitFormBoundary7MA4YWxkTrZu0gW--\r\n";

        InputStream inputStream = new ByteArrayInputStream(request.getBytes());
        HttpRequestParser parser = new HttpRequestParser();

        // when
        HttpRequest httpRequest = parser.parseRequest(inputStream);

        // then
        assertThat(httpRequest)
                .extracting(HttpRequest::getMethod, req -> req.getPath().getBasePath(), HttpRequest::getVersion)
                .containsExactly(HttpMethod.POST, "/upload", HttpVersion.HTTP_1_1);

        assertThat(httpRequest.getHeaders().getHeader(HeaderConstants.CONTENT_TYPE)).isNotNull()
                .contains("multipart/form-data", "boundary=----WebKitFormBoundary7MA4YWxkTrZu0gW");

        byte[] body = httpRequest.getBody().readAllBytes();
        String bodyString = new String(body);
        assertThat(bodyString)
                .contains("This is the content")
                .contains("<binary data of example.png>");
    }

    @DisplayName("urlencoded된 요청이 오는 경우에는 decode된 body를 HttpRequest에 저장한다.")
    @Test
    void urlencoded() throws IOException {
        // given
        String request = """
                POST /user/create HTTP/1.1
                Host: localhost:8080
                Connection: keep-alive
                Content-Length: 93
                Content-Type: application/x-www-form-urlencoded
                
                userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net
                """.replace("\n", "\r\n");

        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(request.getBytes());

        // when
        HttpRequest httpRequest = httpRequestParser.parseRequest(byteArrayInputStream);

        // then
        assertThat(new String(httpRequest.getBody().readAllBytes()))
                .isEqualTo("userId=javajigi&password=password&name=박재성&email=javajigi@slipp.net");
    }

    @DisplayName("urlencoded되어있지 않을 경우 decode하지 않은 body를 HttpRequest에 저장한다.")
    @Test
    void urlencode되어있지_않은_경우() throws IOException {
        // given
        String request = """
                POST /user/create HTTP/1.1
                Host: localhost:8080
                Connection: keep-alive
                Content-Length: 93
                Content-Type: application/json
                
                userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net
                """.replace("\n", "\r\n");
        ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(request.getBytes());


        // when
        HttpRequest httpRequest = httpRequestParser.parseRequest(byteArrayInputStream);
        System.out.println("httpRequest.getBody() = " + httpRequest.getBody());

        // then
        assertThat(new String(httpRequest.getBody().readAllBytes()))
                .isEqualTo("userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net");
    }

    @DisplayName("빈 요청 라인 파싱 시 예외 발생")
    @Test
    void parseRequest_emptyRequestLine() {
        // given
        String rawRequest = "\r\n";
        InputStream inputStream = new ByteArrayInputStream(rawRequest.getBytes());


        // when & then
        assertThatThrownBy(() -> httpRequestParser.parseRequest(inputStream))
                .isInstanceOf(IllegalArgumentException.class)
                .hasMessage("Invalid HTTP request: no end of headers found");
    }

    @DisplayName("정상적인 요청 파싱")
    @Test
    void parseRequest_validRequest() throws Exception {
        // given
        String rawRequest = "GET /test HTTP/1.1\r\n" +
                "Host: localhost\r\n" +
                "Content-Length: 0\r\n\r\n";
        InputStream inputStream = new ByteArrayInputStream(rawRequest.getBytes());


        // when
        HttpRequest request = httpRequestParser.parseRequest(inputStream);

        // then
        assertThat(request).isNotNull()
                .extracting(HttpRequest::getMethod, HttpRequest::getPath, HttpRequest::getVersion, req -> new String(req.getBody().readAllBytes()))
                .containsExactly(HttpMethod.GET, Path.of("/test"), HttpVersion.HTTP_1_1, "");
        assertThat(request.getHeaders().getHeader("Host")).containsExactly("localhost");
    }

    @DisplayName("유효하지 않은 요청 라인 파싱 시 예외 발생")
    @Test
    void parseRequest_invalidRequestLine() {
        // given
        String rawRequest = "INVALID_REQUEST_LINE\r\n";
        InputStream inputStream = new ByteArrayInputStream(rawRequest.getBytes());


        // when & then
        assertThatThrownBy(() -> httpRequestParser.parseRequest(inputStream))
                .isInstanceOf(IllegalArgumentException.class)
                .hasMessage("Invalid HTTP request: no end of headers found");
    }

    @DisplayName("유효하지 않은 헤더 라인 파싱 시 예외 발생")
    @Test
    void parseRequest_invalidHeaderLine() {
        // given
        String rawRequest = "GET /test HTTP/1.1\r\n" +
                "Invalid-Header\r\n\r\n";
        InputStream inputStream = new ByteArrayInputStream(rawRequest.getBytes());


        // when & then
        assertThatThrownBy(() -> httpRequestParser.parseRequest(inputStream))
                .isInstanceOf(BadRequestException.class)
                .hasMessage("Invalid header line: Invalid-Header");
    }

    @DisplayName("유효하지 않은 Content-Length 값 파싱 시 예외 발생")
    @Test
    void parseRequest_invalidContentLength() {
        // given
        String rawRequest = "GET /test HTTP/1.1\r\n" +
                "Host: localhost\r\n" +
                "Content-Length: invalid\r\n\r\n";
        InputStream inputStream = new ByteArrayInputStream(rawRequest.getBytes());


        // when & then
        assertThatThrownBy(() -> httpRequestParser.parseRequest(inputStream))
                .isInstanceOf(NumberFormatException.class);
    }

    @DisplayName("유효하지 않은 Content-Type 값 파싱 시 예외 발생")
    @Test
    void parseRequest_invalidContentType() throws Exception {
        // given
        String rawRequest = "POST /test HTTP/1.1\r\n" +
                "Host: localhost\r\n" +
                "Content-Length: 27\r\n" +
                "Content-Type: application/x-www-form-urlencoded\r\n\r\n" +
                "key1=value1&key2=invalid%XX";
        InputStream inputStream = new ByteArrayInputStream(rawRequest.getBytes());


        // when & then
        assertThatThrownBy(() -> httpRequestParser.parseRequest(inputStream))
                .isInstanceOf(BadRequestException.class)
                .hasMessage("인코딩 에러가 발생했습니다.");
    }
    @DisplayName("헤더가 없는 요청 파싱")
    @Test
    void parseRequest_noHeaders() throws Exception {
        // given
        String rawRequest = "GET /test HTTP/1.1\r\n\r\n";
        InputStream inputStream = new ByteArrayInputStream(rawRequest.getBytes());


        // when
        HttpRequest request = httpRequestParser.parseRequest(inputStream);

        // then
        assertThat(request).isNotNull()
                .extracting(HttpRequest::getMethod, HttpRequest::getPath, HttpRequest::getVersion, req -> new String(req.getBody().readAllBytes()))
                .containsExactly(HttpMethod.GET, Path.of("/test"), HttpVersion.HTTP_1_1, "");
        assertThat(request.getHeaders().getValues()).isEmpty();
    }





}